requirements.txt
pymongo>=4.1.1
requests>=2.28.0
python-dotenv>=0.20.0
tqdm>=4.64.0
